+++ src/interfaces.ts
@@ -1,5 +1,4 @@
 import type { Range, Diagnostic } from 'vscode-languageserver';
-
 import type { Scope } from './Scope';
 import type { BrsFile } from './files/BrsFile';
 import type { XmlFile } from './files/XmlFile';
@@ -10,6 +9,8 @@ import type { ParseMode } from './parser/Parser';
 import type { Program, SourceObj, TranspileObj } from './Program';
 import type { ProgramBuilder } from './ProgramBuilder';
 import type { FunctionStatement } from './parser';
+import type { TranspileState } from './parser/TranspileState';
+import type { SourceNode } from 'source-map';
 
 export interface BsDiagnostic extends Diagnostic {
     file: File;
@@ -159,7 +160,7 @@ export interface CallableContainer {
     scope: Scope;
 }
 
-export type CallableContainerMap = Record<string, CallableContainer[]>;
+export type CallableContainerMap = Map<string, CallableContainer[]>;
 
 export interface CommentFlag {
     file: BrsFile;
@@ -200,4 +201,8 @@ export interface CompilerPlugin {
     afterFileTranspile?: (entry: TranspileObj) => void;
     beforeFileDispose?: (file: BscFile) => void;
     afterFileDispose?: (file: BscFile) => void;
+}
+
+export interface TypedefProvider {
+    getTypedef(state: TranspileState): Array<SourceNode | string>;
 }