+++ src/parser/Statement.ts
@@ -1,17 +1,18 @@
 /* eslint-disable no-bitwise */
 import type { Token, Identifier } from '../lexer';
-import { TokenKind, CompoundAssignmentOperators } from '../lexer';
+import { CompoundAssignmentOperators, TokenKind } from '../lexer';
 import { SourceNode } from 'source-map';
-import type { BinaryExpression, Expression, NamespacedVariableNameExpression, FunctionExpression } from './Expression';
+import type { BinaryExpression, Expression, NamespacedVariableNameExpression, FunctionExpression, AnnotationExpression } from './Expression';
 import { CallExpression, VariableExpression, LiteralExpression } from './Expression';
 import { util } from '../util';
 import { Range, Position } from 'vscode-languageserver';
 import type { TranspileState } from './TranspileState';
 import { ParseMode, Parser } from './Parser';
 import type { WalkVisitor, WalkOptions } from '../astUtils/visitors';
-import { walk, InternalWalkMode, createVisitor, WalkMode } from '../astUtils/visitors';
-import { isCallExpression, isClassFieldStatement, isClassMethodStatement, isCommentStatement, isExpression, isExpressionStatement, isFunctionStatement } from '../astUtils/reflection';
-import { BrsInvalid } from '../brsTypes/BrsType';
+import { InternalWalkMode, walk, createVisitor, WalkMode } from '../astUtils/visitors';
+import { isCallExpression, isClassFieldStatement, isClassMethodStatement, isCommentStatement, isExpression, isExpressionStatement, isFunctionStatement, isLiteralExpression } from '../astUtils/reflection';
+import { BrsInvalid, ValueKind, valueKindFromString, valueKindToString } from '../brsTypes/BrsType';
+import type { TypedefProvider } from '../interfaces';
 
 /**
  * A BrightScript statement
@@ -22,6 +23,11 @@ export abstract class Statement {
      *  The starting and ending location of the statement.
      **/
     public abstract range: Range;
+
+    /**
+     * Statement annotations
+     */
+    public annotations: AnnotationExpression[];
 
     public abstract transpile(state: TranspileState): Array<SourceNode | string>;
 
@@ -51,7 +57,10 @@ export class EmptyStatement extends Statement {
     }
 }
 
-export class Body extends Statement {
+/**
+ * This is a top-level statement. Consider this the root of the AST
+ */
+export class Body extends Statement implements TypedefProvider {
     constructor(
         public statements: Statement[] = []
     ) {
@@ -94,6 +103,20 @@ export class Body extends Statement {
             }
 
             result.push(...statement.transpile(state));
+        }
+        return result;
+    }
+
+    getTypedef(state: TranspileState) {
+        let result = [];
+        for (const statement of this.statements) {
+            //if the current statement supports generating typedef, call it
+            if ('getTypedef' in statement) {
+                result.push(
+                    ...(statement as TypedefProvider).getTypedef(state),
+                    state.newline()
+                );
+            }
         }
         return result;
     }
@@ -221,7 +244,7 @@ export class ExpressionStatement extends Statement {
     }
 }
 
-export class CommentStatement extends Statement implements Expression {
+export class CommentStatement extends Statement implements Expression, TypedefProvider {
     constructor(
         public comments: Token[]
     ) {
@@ -258,6 +281,10 @@ export class CommentStatement extends Statement implements Expression {
             }
         }
         return result;
+    }
+
+    public getTypedef(state: TranspileState) {
+        return this.transpile(state);
     }
 
     walk(visitor: WalkVisitor, options: WalkOptions) {
@@ -312,7 +339,7 @@ export class ExitWhileStatement extends Statement {
     }
 }
 
-export class FunctionStatement extends Statement {
+export class FunctionStatement extends Statement implements TypedefProvider {
     constructor(
         public name: Identifier,
         public func: FunctionExpression,
@@ -346,6 +373,10 @@ export class FunctionStatement extends Statement {
         };
 
         return this.func.transpile(state, nameToken);
+    }
+
+    getTypedef(state: TranspileState) {
+        return this.func.getTypedef(state, this.name);
     }
 
     walk(visitor: WalkVisitor, options: WalkOptions) {
@@ -994,7 +1025,7 @@ export class IndexedSetStatement extends Statement {
     }
 }
 
-export class LibraryStatement extends Statement {
+export class LibraryStatement extends Statement implements TypedefProvider {
     constructor(
         readonly tokens: {
             library: Token;
@@ -1025,12 +1056,16 @@ export class LibraryStatement extends Statement {
         return result;
     }
 
+    getTypedef(state: TranspileState) {
+        return this.transpile(state);
+    }
+
     walk(visitor: WalkVisitor, options: WalkOptions) {
         //nothing to walk
     }
 }
 
-export class NamespaceStatement extends Statement {
+export class NamespaceStatement extends Statement implements TypedefProvider {
     constructor(
         public keyword: Token,
         //this should technically only be a VariableExpression or DottedGetExpression, but that can be enforced elsewhere
@@ -1063,6 +1098,25 @@ export class NamespaceStatement extends Statement {
         return this.body.transpile(state);
     }
 
+    getTypedef(state: TranspileState) {
+        let result = [
+            'namespace ',
+            ...this.nameExpression.getName(ParseMode.BrighterScript),
+            state.newline()
+        ];
+        state.blockDepth++;
+        result.push(
+            state.indent(),
+            ...this.body.getTypedef(state)
+        );
+        state.blockDepth--;
+
+        result.push(
+            'end namespace'
+        );
+        return result;
+    }
+
     walk(visitor: WalkVisitor, options: WalkOptions) {
         if (options.walkMode & InternalWalkMode.walkExpressions) {
             walk(this, 'nameExpression', visitor, options);
@@ -1073,7 +1127,7 @@ export class NamespaceStatement extends Statement {
     }
 }
 
-export class ImportStatement extends Statement {
+export class ImportStatement extends Statement implements TypedefProvider {
     constructor(
         readonly importToken: Token,
         readonly filePathToken: Token
@@ -1111,12 +1165,24 @@ export class ImportStatement extends Statement {
         ];
     }
 
+    /**
+     * Get the typedef for this statement
+     */
+    public getTypedef(state: TranspileState) {
+        return [
+            this.importToken.text,
+            ' ',
+            //replace any `.bs` extension with `.brs`
+            this.filePathToken.text.replace(/\.bs"?$/i, '.brs"')
+        ];
+    }
+
     walk(visitor: WalkVisitor, options: WalkOptions) {
         //nothing to walk
     }
 }
 
-export class ClassStatement extends Statement {
+export class ClassStatement extends Statement implements TypedefProvider {
 
     constructor(
         readonly classKeyword: Token,
@@ -1178,6 +1244,36 @@ export class ClassStatement extends Statement {
         );
         //make the class assembler (i.e. the public-facing class creator method)
         result.push(...this.getTranspiledClassFunction(state));
+        return result;
+    }
+
+    getTypedef(state: TranspileState) {
+        const result = [] as Array<string | SourceNode>;
+        result.push(
+            'class ',
+            this.name.text
+        );
+        if (this.extendsKeyword && this.parentClassName) {
+            result.push(
+                ` extends ${this.parentClassName.getName(ParseMode.BrighterScript)}`
+            );
+        }
+        result.push(state.newline());
+        state.blockDepth++;
+        for (const member of this.body) {
+            if ('getTypedef' in member) {
+                result.push(
+                    state.indent(),
+                    ...(member as TypedefProvider).getTypedef(state),
+                    state.newline()
+                );
+            }
+        }
+        state.blockDepth--;
+        result.push(
+            state.indent(),
+            'end class'
+        );
         return result;
     }
 
@@ -1495,6 +1591,20 @@ export class ClassMethodStatement extends FunctionStatement {
         return this.func.transpile(state);
     }
 
+    getTypedef(state: TranspileState) {
+        const result = [] as string[];
+        if (this.accessModifier) {
+            result.push(
+                this.accessModifier.text,
+                ' '
+            );
+        }
+        result.push(
+            ...super.getTypedef(state)
+        );
+        return result;
+    }
+
     /**
      * All child classes must call the parent constructor. The type checker will warn users when they don't call it in their own class,
      * but we still need to call it even if they have omitted it. This injects the super call if it's missing
@@ -1598,7 +1708,7 @@ export class ClassMethodStatement extends FunctionStatement {
     }
 }
 
-export class ClassFieldStatement extends Statement {
+export class ClassFieldStatement extends Statement implements TypedefProvider {
 
     constructor(
         readonly accessModifier?: Token,
@@ -1615,10 +1725,38 @@ export class ClassFieldStatement extends Statement {
         );
     }
 
+    /**
+     * Derive a ValueKind from the type token, or the intial value.
+     * Defaults to `ValueKind.Dynamic`
+     */
+    private getType() {
+        if (this.type) {
+            return valueKindFromString(this.type.text);
+        } else if (isLiteralExpression(this.initialValue)) {
+            return this.initialValue.value.kind;
+        } else {
+            return ValueKind.Dynamic;
+        }
+    }
+
     public readonly range: Range;
 
     transpile(state: TranspileState): Array<SourceNode | string> {
         throw new Error('transpile not implemented for ' + Object.getPrototypeOf(this).constructor.name);
+    }
+
+    getTypedef(state: TranspileState) {
+        const result = [];
+        if (this.name) {
+            result.push(
+                this.accessModifier?.text ?? 'public',
+                ' ',
+                this.name?.text,
+                ' as ',
+                valueKindToString(this.getType()).toLowerCase()
+            );
+        }
+        return result;
     }
 
     walk(visitor: WalkVisitor, options: WalkOptions) {